updates it to run on the server
